@@ -2,9 +2,9 @@
 
 buildscript {
     ext {
-        agp_version = '8.9.0-alpha05'
+        agp_version = '8.11.0-alpha05'
     }
-    ext.kotlin_version = '2.0.20'
+    ext.kotlin_version = '2.1.20'
     repositories {
         maven { url 'https://jitpack.io' }
         google()
